+++ R/dss--main.R
@@ -2,6 +2,7 @@
 #' @importFrom shiny icon
 
 dss_ui <- function() { bslib::page_navbar(
+  id = "navbar",
   title = dss_title(),
   header = dss_header(),
   footer = dss_footer(),
@@ -58,30 +59,21 @@ dss_ui <- function() { bslib::page_navbar(
 dss_server <- function(input, output, session) {
   clicked_cluster_id <- shiny::reactive({ input$map_shape_click$id })
 
+  run <- shiny::reactiveVal(1)
+  shiny::observe(run(run() + 1)) |> shiny::bindEvent(input$run)
+
   parameters <- dss_input_server("dss_input")
-  param_hash <- shiny::reactive( digest::digest(lapply(parameters, \(r) r())) )
-  shiny::observe(shiny::updateActionButton(session, "run", disabled = FALSE)) |>
-    shiny::bindEvent(param_hash())
-
-  sim <- dss_run_server("dss_run", parameters = parameters, run = input$run)
-  shiny::observe(shiny::updateActionButton(session, "run", disabled = TRUE)) |>
-    shiny::bindEvent(sim())
-
+  sim <- dss_run_server("dss_run", parameters = parameters, run = run)
   dss_output_server("dss_output",
                     simulation = sim,
                     clicked_cluster_id = clicked_cluster_id)
 
-  vmap <- shiny::reactive(
-    get_layer_aux(sim(), "inp")[["cluster_variety_map"]]
-    )
-
+  vmap <- shiny::reactive(get_layer_aux(sim(), "inp")[["cluster_variety_map"]])
   output$map <- leaflet::renderLeaflet(
     plot_albufera_clusters(cluster_variety_map = vmap())
     ) |>
     shiny::snapshotExclude()
-
   shiny::observeEvent(input$hide_map_card, shinyjs::hide("map_card"))
-
   shiny::observeEvent(input$show_map_card, shinyjs::show("map_card"))
 
   shiny::observeEvent(input$show_about, shiny::showModal(dss_about_modal()))