+++ NEWS.md
@@ -1,3 +1,29 @@
+# erahumed 0.16.1
+
+### GUI
+
+* The X ranges of time series plots in the Output section are now synchronized
+(#231).
+
+* Parameter updates trigger a notification that prompts the user to re-run the
+simulation (#301).
+
+* Date range input is now a `shinyWidgets::airDatepickerInput()` (#269).
+
+* Run button functionality restored (#309).
+
+* A new `selectInput()` allows to choose the type of risk to be plotted 
+(acute *vs.* chronic).
+
+* A new radio button allows to choose whether to show mass or density in 
+exposure plots.
+
+* Card headers in the Output section are now colored in black.
+
+### Dependencies
+
+* We start importing `{shinyWidgets}`.
+
 # erahumed 0.16.0
 
 This version provides the first implementation of risk assessment through SSDs.
@@ -15,7 +41,6 @@ outputs for risk.
 
 * `cluster_id` and `ditch` columns of the respective `ctc` and `ctd` layers have
 been renamed to `element_id`.
->>>>>>> f6a06c9c485298b85ca58bc590f54785529011ce
 
 # erahumed 0.15.1